[BLINK-16766195] [table] Fix performance regression of count distinct
Summary:
fix T16766195

We optimize distinct aggregate in previous commit, that is when we found the input stream is append only, we will not record the occur times of the distinct key, and only put the one time when it first been seen.
But this result in more cache miss and the get-latency increase.

Reviewers: 玉兆

Reviewed By: 玉兆

Differential Revision: https://aone.alibaba-inc.com/code/D705590