[BLINK-17978731] bridge ExternalCatalogTable to new catalog APIs
Summary:
As discussed with Xuefu, we should reuse whatever Blink uses right now to move forward.

In this PR, we will temporarily not use the new CatalogTable. We will use the old ExternalCatalogTable for now in the new catalog APIs, in order to make the whole TableSource/Sink registration workflow work.

The whole registration workflow is successful now and can be verified in CatalogCalciteSchemaTest.java.testGetTable()

Reviewers: xuefu.z

Reviewed By: xuefu.z

Subscribers: P710480

Differential Revision: https://aone.alibaba-inc.com/code/D780744