[BLINK-18570046] remove old ExternalCatalog related code from TableEnvironment
Summary:
Completely remove old ExternalCatalog related legacy code from TableEnvironment to avoid confusing users

It may be worth discussing whether or not it's really necessary to expose getTableStats() and alterTableStats() in TableEnvironment. These methods are only introduced in Blink, but not in Flink community version.

Shall we at least hide those APIs or mark them as experimental for now? Currently I'm not able to remove getTableStats() and alterTableStats() from TableEnvironment because they are widely used in many unit tests.

Reviewers: 乔燃, 鲁尼, 青陆, 玉兆, 大沙, xuefu.z

Reviewed By: xuefu.z

Subscribers: P710480

Differential Revision: https://aone.alibaba-inc.com/code/D815129