BLINK-18739887: flink read hive date type deal not correct
Summary:
Encoding java.sql.Date to integer was done in a wrong way such that when a Date instance is converted
to an integer and then converted back to Date, some precision is lost during the conversion. The fix changed
the encoding, following how it's done in projects like Hive and ORC.

Test Plan: Manually verified.

Reviewers: 乔燃, 鲁尼, bowen.li

Reviewed By: bowen.li

Subscribers: P710480

Differential Revision: https://aone.alibaba-inc.com/code/D827174